Battle::Bridge: inline some methods
